+++ app/services/stripe_service.rb
@@ -50,7 +50,7 @@ class StripeService
       if coupon.type == 'percent_off'
         stp_coupon[:percent_off] = coupon.percent_off
       elsif coupon.type == 'amount_off'
-        stp_coupon[:amount_off] = coupon.amount_off
+        stp_coupon[:amount_off] = stripe_amount(coupon.amount_off)
         stp_coupon[:currency] = Setting.get('stripe_currency')
       end
 
@@ -59,6 +59,13 @@ class StripeService
       stp_coupon[:max_redemptions] = coupon.max_usages unless coupon.max_usages.nil?
 
       Stripe::Coupon.create(stp_coupon, api_key: Setting.get('stripe_secret_key'))
+    end
+
+    def stripe_amount(amount)
+      currency = Setting.get('stripe_currency')
+      return amount / 100 if zero_decimal_currencies.any? { |s| s.casecmp(currency).zero? }
+
+      amount
     end
 
     private
@@ -91,7 +98,7 @@ class StripeService
 
     def create_price(amount, stp_product_id, name, monthly: false)
       params = {
-        unit_amount: amount,
+        unit_amount: stripe_amount(amount),
         currency: Setting.get('stripe_currency'),
         product: stp_product_id,
         nickname: name
@@ -106,6 +113,11 @@ class StripeService
 
       customer_id = payment_schedule.invoicing_profile.user.stp_customer_id
       Stripe::Customer.update(customer_id, { balance: -payment_schedule.wallet_amount }, { api_key: Setting.get('stripe_secret_key') })
+    end
+
+    # @see https://stripe.com/docs/currencies#zero-decimal
+    def zero_decimal_currencies
+      %w[BIF CLP DJF GNF JPY KMF KRW MGA PYG RWF UGX VND VUV XAF XOF XPF]
     end
   end
 end